rename stream_ml to stream_mapper
