+++ .pre-commit-config.yaml
@@ -57,7 +57,7 @@ repos:
         additional_dependencies:
           - numpy>=1.20,<1.24
           - typing_extensions>=4.4.0
-          - "stream_ml.core @ git+https://github.com/nstarman/stream_ml-core@main#egg=stream_ml-core"
+          - "stream_mapper.core @ git+https://github.com/GalOrrery/stream_mapper-core@main#egg=stream_mapper-core"
         args:
           - --strict
           - --allow-subclassing-any
