+++ src/stream_mapper/pytorch/builtin/compat/__init__.py
@@ -1,0 +1,7 @@
+"""Stream Memberships Likelihood, with ML."""
+
+from stream_mapper.pytorch.builtin.compat.kde import KDEModel
+from stream_mapper.pytorch.builtin.compat.nflow import NFlowModel
+from stream_mapper.pytorch.builtin.compat.zuko import ZukoFlowModel
+
+__all__ = ("ZukoFlowModel", "NFlowModel", "KDEModel")
