+++ src/stream_mapper/pytorch/nn.py
