+++ src/stream_mapper/pytorch/params/scaler.py
@@ -1,0 +1,6 @@
+"""Core feature."""
+
+from stream_mapper.core.params import scaler
+from stream_mapper.core.params.scaler import *  # noqa: F403
+
+__all__ = scaler.__all__
