+++ src/stream_mapper/pytorch/typing.py
@@ -1,0 +1,8 @@
+"""Pytorch type hints."""
+
+__all__ = ("Array", "NNModel", "ArrayNamespace")
+
+from torch import Tensor as Array
+from torch.nn import Module as NNModel
+
+from stream_mapper.core.typing import ArrayNamespace
