+++ src/stream_mapper/pytorch/utils.py
@@ -1,0 +1,6 @@
+"""Utilities."""
+
+from stream_mapper.core import utils
+from stream_mapper.core.utils import *  # noqa: F403
+
+__all__ = utils.__all__
