--- src/stream_ml/pytorch/builtin/compat/__init__.py
@@ -1,7 +1,0 @@
-"""Stream Memberships Likelihood, with ML."""
-
-from stream_ml.pytorch.builtin.compat.kde import KDEModel
-from stream_ml.pytorch.builtin.compat.nflow import NFlowModel
-from stream_ml.pytorch.builtin.compat.zuko import ZukoFlowModel
-
-__all__ = ("ZukoFlowModel", "NFlowModel", "KDEModel")
