--- src/stream_ml/pytorch/params/__init__.py
@@ -1,30 +1,0 @@
-"""Parameters."""
-
-__all__: list[str] = [
-    # modules
-    "bounds",
-    "scaler",
-    # parameters
-    "ModelParameter",
-    "ModelParameters",
-    "ModelParametersField",
-    # values
-    "Params",
-    "freeze_params",
-    "unfreeze_params",
-    "set_param",
-    "add_prefix",
-]
-
-from stream_ml.core.params import (
-    ModelParameter,
-    ModelParameters,
-    ModelParametersField,
-    Params,
-    add_prefix,
-    freeze_params,
-    set_param,
-    unfreeze_params,
-)
-
-from stream_ml.pytorch.params import bounds, scaler
