--- src/stream_ml/pytorch/params/scaler.py
@@ -1,6 +1,0 @@
-"""Core feature."""
-
-from stream_ml.core.params import scaler
-from stream_ml.core.params.scaler import *  # noqa: F403
-
-__all__ = scaler.__all__
