--- src/stream_ml/pytorch/utils.py
@@ -1,6 +1,0 @@
-"""Utilities."""
-
-from stream_ml.core import utils
-from stream_ml.core.utils import *  # noqa: F403
-
-__all__ = utils.__all__
