+++ tox.ini
@@ -66,8 +66,8 @@ extras =
 
 commands =
     pip freeze
-    !cov-!double: pytest --pyargs {toxinidir}/tests stream_ml {posargs}
-    cov-!double: pytest --pyargs {toxinidir}/tests stream_ml --cov stream_ml --cov-config={toxinidir}/pyproject.toml {posargs}
+    !cov-!double: pytest --pyargs {toxinidir}/tests stream_mapper {posargs}
+    cov-!double: pytest --pyargs {toxinidir}/tests stream_mapper --cov stream_mapper --cov-config={toxinidir}/pyproject.toml {posargs}
     cov: coverage xml -o {toxinidir}/coverage.xml
 
 [testenv:codestyle]
@@ -82,7 +82,7 @@ commands =
 [pytest]
 testpaths =
     tests
-    src/stream_ml
+    src/stream_mapper
 astropy_header = True
 doctest_plus = enabled
 text_file_format = rst
